Did Day 8 and ) and Updates
We no directly get the input from AOC using an envirentmental var to run this. this will make the github action harder to run however.

Also did day 8 and 9 this comes with mor eutlity classes and updated model classes
