+++ AOCFramework/build.gradle
@@ -16,6 +16,7 @@ java {
 }
 
 dependencies {
+    testImplementation('org.assertj:assertj-core:3.24.2')
     testImplementation platform('org.junit:junit-bom:5.9.1')
     testImplementation 'org.junit.jupiter:junit-jupiter'
 }
