--- AOCFramework/src/main/java/aoc/framework/DaySolution.java
@@ -1,11 +1,0 @@
-package aoc.framework;
-
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
-
-@Retention(RetentionPolicy.RUNTIME)
-public @interface DaySolution {
-    int year();
-
-    int day();
-}
