+++ AOCFramework/src/main/java/aoc/framework/solution/Solution.java
@@ -1,0 +1,11 @@
+package aoc.framework.solution;
+
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+
+@Retention(RetentionPolicy.RUNTIME)
+public @interface Solution {
+    int year();
+
+    int day();
+}
