--- AOCSolutions/src/main/resources/202301/input_short_part_2.txt
@@ -1,7 +1,0 @@
-two1nine
-eightwothree
-abcone2threexyz
-xtwone3four
-4nineeightseven2
-zoneight234
-7pqrstsixteen
