--- AOCSolutions/src/main/resources/202303/input_short_part_1.txt
@@ -1,10 +1,0 @@
-467..114..
-...*......
-..35..633.
-......#.+.
-617*.....1
-.....+.58.
-..592.....
-......755.
-...$.*....
-.664.598..
