--- AOCSolutions/src/main/resources/202306/input.txt
@@ -1,2 +1,0 @@
-Time:        56     97     78     75
-Distance:   546   1927   1131   1139
