--- AOCSolutions/src/main/resources/202306/input_short_part_1.txt
@@ -1,2 +1,0 @@
-Time:      7  15   30
-Distance:  9  40  200
