--- AOCSolutions/src/main/resources/202307/input_short_part1.txt
@@ -1,5 +1,0 @@
-32T3K 765
-T55J5 684
-KK677 28
-KTJJT 220
-QQQJA 483
