--- AOCSolutions/src/main/resources/202308/input_part_1_repeating.txt
@@ -1,5 +1,0 @@
-LLR
-
-AAA = (BBB, BBB)
-BBB = (AAA, ZZZ)
-ZZZ = (ZZZ, ZZZ)
