+++ src/main/java/aoc/AdventOfCode2022Solutions.java
@@ -1,8 +1,8 @@
 package aoc;
 
 import aoc.factory.DaySolutionFactory;
-import aoc.framework.Day;
-import aoc.framework.DaySolution;
+import aoc.framework.solution.AOCSolution;
+import aoc.framework.solution.Solution;
 
 import java.util.Comparator;
 import java.util.List;
@@ -22,9 +22,9 @@ public class AdventOfCode2022Solutions {
     private static void solveAll() {
         for (int year : getAllYears()) {
             System.out.printf("Solutions for %s Advent Of Code %n", year);
-            List<? extends Day> days = getAllSolutionsForYear(year);
+            List<? extends AOCSolution> days = getAllSolutionsForYear(year);
             for (int i = 0; i < days.size(); i++) {
-                System.out.printf("\t >Day %s %n", days.get(i).getClass().getAnnotation(DaySolution.class).day());
+                System.out.printf("\t >Day %s %n", days.get(i).getClass().getAnnotation(Solution.class).day());
                 long startTime = System.currentTimeMillis();
                 String solutionPart1 = days.get(i).solvePart1();
                 long endTime = System.currentTimeMillis();
@@ -39,21 +39,21 @@ public class AdventOfCode2022Solutions {
             }
         }
     }
-    
+
     private static List<Integer> getAllYears() {
         return DaySolutionFactory.createDaySolutions().stream()
-                .filter(day -> Objects.nonNull(day.getClass().getAnnotation(DaySolution.class)))
-                .map(day -> day.getClass().getAnnotation(DaySolution.class).year())
+                .filter(day -> Objects.nonNull(day.getClass().getAnnotation(Solution.class)))
+                .map(day -> day.getClass().getAnnotation(Solution.class).year())
                 .distinct()
                 .sorted()
                 .toList();
     }
 
-    private static List<? extends Day> getAllSolutionsForYear(int year) {
+    private static List<? extends AOCSolution> getAllSolutionsForYear(int year) {
         return DaySolutionFactory.createDaySolutions().stream()
-                .filter(day -> day.getClass().getAnnotation(DaySolution.class) != null)
-                .filter(day -> day.getClass().getAnnotation(DaySolution.class).year() == year)
-                .sorted(Comparator.comparingInt((Day o) -> o.getClass().getAnnotation(DaySolution.class).day()))
+                .filter(day -> day.getClass().getAnnotation(Solution.class) != null)
+                .filter(day -> day.getClass().getAnnotation(Solution.class).year() == year)
+                .sorted(Comparator.comparingInt((AOCSolution o) -> o.getClass().getAnnotation(Solution.class).day()))
                 .toList();
     }
